move dummy executor out to wiggle:: for reuse
